last code

+++ backend/controllers/projectController.js
@@ -2,7 +2,7 @@ const { Project } = require('../models')
 
 const getProjects = async (req, res) => {
     try {
-        const projects = await Project.find().sort({ createdAt: -1 })
+        const projects = await Project.find({ owner: req.user.id }).sort({ createdAt: -1 })
         return res.status(200).json(projects)
     } catch (error) {
         console.error(error)
@@ -19,9 +19,9 @@ const createProject = async (req, res) =>{
 
     try {
         const project = await Project.create({
-            // owner: req.user.id,
+            owner: req.user.id,
             title: title.trim(),
-            duaDate: dueDate ||"",
+            dueDate: dueDate ||"",
             tasks: (tasks || []).map((text) => ({text}))
         })
 
@@ -36,7 +36,7 @@ const deletePorject = async (req, res) => {
     const {projectId} = req.params
 
     try {
-        const project = await Project.findOneAndDelete({_id: projectId, })
+        const project = await Project.findOneAndDelete({_id: projectId, owner: req.user.id})
         if(!project){
             return res.status(404).json({error: "Project not found"})
         }
